support serial range and names, refs #6852
If /server/@cleanup is set to "yes" in the server.xml, it is now
possible to set /server/users/@temporary_serial_min and _max in addition
to the /server/users/@temporary regular expression.

If a player's serial is between those serials OR it's name matches the
regex, he is regarded a temporary player.